+++ core/server/server_impl.hpp
@@ -6,7 +6,6 @@
 #include "../network/tcp/connection.hpp"
 
 #include <string>
-#include <map>
 
 namespace Lambda::Server {
 
@@ -14,7 +13,7 @@ namespace Lambda::Server {
 		HTTP, WS, SSE
 	};
 
-	void connectionHandler(Network::TCP::Connection&& conn, const ServeOptions& config, const ServerlessCallback& handlerCallback) noexcept;
+	void connectionHandler(Network::TCP::Connection&& conn, const ServeOptions& config, const RequestCallback& handlerCallback) noexcept;
 
 	namespace Pages {
 		HTTP::Response renderErrorPage(HTTP::Status code, const std::string& message, ErrorResponseType type);
